+++ docs/service-worker.js
@@ -14,7 +14,7 @@
 importScripts("https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js");
 
 importScripts(
-  "./precache-manifest.9aef8b5b17a6286950f9d5c8a1d7ca97.js"
+  "./precache-manifest.fcf58e96cb5abd3d431488a76d91bba3.js"
 );
 
 self.addEventListener('message', (event) => {