+++ src/components/Header/Header.js
@@ -22,6 +22,7 @@ import classNames from "classnames";
 
 //images
 import profile from "../../images/main-profile.png";
+import config from "../../config";
 
 // styles
 import useStyles from "./styles";
@@ -37,6 +38,12 @@ import {
   useLayoutDispatch,
   toggleSidebar
 } from "../../context/LayoutContext";
+import {
+  useManagementDispatch,
+  useManagementState,
+} from '../../context/ManagementContext'
+
+import { actions } from '../../context/ManagementContext'
 import { useUserDispatch, signOut } from "../../context/UserContext";
 
 const messages = [
@@ -100,6 +107,7 @@ export default function Header(props) {
   var layoutState = useLayoutState();
   var layoutDispatch = useLayoutDispatch();
   var userDispatch = useUserDispatch();
+  const managementDispatch = useManagementDispatch();
 
   // local
   var [mailMenu, setMailMenu] = useState(null);
@@ -109,6 +117,12 @@ export default function Header(props) {
   var [profileMenu, setProfileMenu] = useState(null);
   var [isSearchOpen, setSearchOpen] = useState(false);
   const [isSmall, setSmall] = useState(false);
+
+  const managementValue = useManagementState()
+
+  useEffect(() => {
+    actions.doFind(sessionStorage.getItem('user_id'))(managementDispatch)
+  }, [])
 
   useEffect(function() {
     window.addEventListener("resize", handleWindowWidthChange);
@@ -219,7 +233,7 @@ export default function Header(props) {
         >
           <Avatar
             alt="Robert Cotton"
-            src={profile}
+            src={config.isBackend ? (managementValue.currentUser && managementValue.currentUser.avatar.length >=1 && managementValue.currentUser.avatar[managementValue.currentUser.avatar.length-1].publicUrl || profile) : profile}
             classes={{ root: classes.headerIcon }}
           />
         </IconButton>