+++ src/components/Layout/Layout.js
@@ -11,19 +11,8 @@ import {
 import {
     Fab,
     IconButton,
-    Box,
-    Grid,
-    Breadcrumbs,
-    Tabs,
     Tab,
 } from '@material-ui/core'
-import {
-    NavigateNext as NavigateNextIcon,
-    CalendarToday as CalendarIcon,
-    ChatBubbleOutline as ChatIcon,
-    AddShoppingCart as AddIcon,
-    StarBorder as StarIcon,
-} from '@material-ui/icons'
 import { withStyles } from '@material-ui/styles'
 
 // styles
@@ -33,8 +22,7 @@ import useStyles from './styles'
 import Header from '../Header'
 import Sidebar from '../Sidebar'
 import Footer from '../Footer'
-import Widget from '../Widget'
-import { Link, Typography, Button } from '../../components/Wrappers'
+import { Link } from '../../components/Wrappers'
 import ColorChangeThemePopper from './components/ColorChangeThemePopper'
 
 // pages
@@ -75,6 +63,7 @@ import Calendar from '../../pages/calendar'
 import UserList from '../../pages/user'
 import UserAdd from '../../pages/user/AddUser'
 import UserEdit from '../../pages/user/EditUser'
+import BreadCrumbs from '../../components/BreadCrumbs'
 
 // context
 import { useLayoutState } from '../../context/LayoutContext'
@@ -82,16 +71,6 @@ import { ProductsProvider } from '../../context/ProductContext'
 
 //Sidebar structure
 import structure from '../Sidebar/SidebarStructure'
-
-// Tab styling
-
-const CustomTab = withStyles(theme => ({
-    root: {
-        minWidth: 72,
-        textTransform: 'none',
-        fontWeight: 400,
-    },
-}))(props => <Tab {...props} />)
 
 function Layout(props) {
     const classes = useStyles()
@@ -128,146 +107,7 @@ function Layout(props) {
                 })}
             >
                 <div className={classes.fakeToolbar} />
-                <Widget
-                    disableWidgetMenu
-                    inheritHeight
-                    className={classes.margin}
-                    bodyClass={classes.navPadding}
-                >
-                    <Grid
-                        container
-                        direction="row"
-                        justify="space-between"
-                        alignItems="center"
-                        wrap={'nowrap'}
-                        style={{ overflowX: 'auto' }}
-                    >
-                        {structure.map(c => {
-                            if (
-                                !c.children &&
-                                window.location.hash.includes(c.link) &&
-                                c.link
-                            ) {
-                                return (
-                                    <Box
-                                        display="flex"
-                                        alignItems="center"
-                                        key={c.id}
-                                    >
-                                        <Breadcrumbs aria-label="breadcrumb">
-                                            <Typography variant="h4">
-                                                {c.label}
-                                            </Typography>
-                                        </Breadcrumbs>
-                                        {window.location.hash.includes(
-                                            '/app/dashboard'
-                                        ) && (
-                                            <Tabs
-                                                value={value}
-                                                onChange={handleChange}
-                                                aria-label="simple tabs example"
-                                                variant="scrollable"
-                                                scrollButtons="auto"
-                                                style={{ marginLeft: 38 }}
-                                            >
-                                                <CustomTab
-                                                    label="Today"
-                                                    {...a11yProps(0)}
-                                                />
-                                                <CustomTab
-                                                    label="This week"
-                                                    {...a11yProps(1)}
-                                                />
-                                                <CustomTab
-                                                    label="This month"
-                                                    {...a11yProps(2)}
-                                                />
-                                                <CustomTab
-                                                    label="This year"
-                                                    {...a11yProps(3)}
-                                                />
-                                            </Tabs>
-                                        )}
-                                    </Box>
-                                )
-                            } else if (c.children) {
-                                return c.children.map(currentInner => {
-                                    if (
-                                        window.location.hash.includes(
-                                            currentInner.link
-                                        )
-                                    ) {
-                                        return (
-                                            <Breadcrumbs
-                                                separator={
-                                                    <NavigateNextIcon fontSize="small" />
-                                                }
-                                                aria-label="breadcrumb"
-                                                key={c.id}
-                                            >
-                                                <Typography variant={'h6'}>
-                                                    {c.label}
-                                                </Typography>
-                                                <Typography
-                                                    variant={'h6'}
-                                                    color="primary"
-                                                >
-                                                    {currentInner.label}
-                                                </Typography>
-                                            </Breadcrumbs>
-                                        )
-                                    } else {
-                                        return null
-                                    }
-                                })
-                            } else {
-                                return null
-                            }
-                        })}
-                        {window.location.hash.includes('/app/dashboard') && (
-                            <Box display="flex" alignItems="center">
-                                <CalendarIcon
-                                    className={classes.calendarIcon}
-                                />
-                                <Typography className={classes.date}>
-                                    29 Oct 2019, Tuesday
-                                </Typography>
-                                <Button
-                                    variant="contained"
-                                    color="secondary"
-                                    className={classes.button}
-                                >
-                                    Latest Reports
-                                </Button>
-                            </Box>
-                        )}
-                        {window.location.hash.includes('/app/ecommerce') && (
-                            <Box display="flex" alignItems="center">
-                                <Box>
-                                    <IconButton aria-label="chat">
-                                        <ChatIcon
-                                            className={classes.ecommerceIcon}
-                                        />
-                                    </IconButton>
-                                </Box>
-                                <Box>
-                                    <IconButton aria-label="add_to_cart">
-                                        <AddIcon
-                                            className={classes.ecommerceIcon}
-                                        />
-                                    </IconButton>
-                                </Box>
-                                <Box>
-                                    <IconButton aria-label="rate">
-                                        <StarIcon
-                                            className={classes.ecommerceIcon}
-                                        />
-                                    </IconButton>
-                                </Box>
-                            </Box>
-                        )}
-                    </Grid>
-                </Widget>
+                <BreadCrumbs />
                 <Switch>
                     <Route path="/app/dashboard" component={Dashboard} />
                     <Route
@@ -343,7 +183,7 @@ function Layout(props) {
                         exact
                         path="/app/ui"
                         render={() => <Redirect to="/app/ui/icons" />}
-                    />]
+                    />
                     <Route
                         exact
                         path="/app/core"
@@ -387,7 +227,8 @@ function Layout(props) {
                     />
                     <Route path="/app/user/list" component={UserList} />
                     <Route path="/app/user/add" component={UserAdd} />
-                    <Route path="/app/user/edit" component={UserEdit} />
+                    <Route path="/app/user/:id/edit" component={UserEdit} />
+                    <Route path="/app/user/:id" component={UserEdit} />
                 </Switch>
                 <Fab
                     color="primary"